Improve generated definition for some types of expression.
When visiting the class expression that logically defines a term, keep
track of the type of expression we are visiting (ObjectIntersectionOf or
ObjectSomeValuesOf), so that when we inject the label of a class, we can
insert a connecting "is a(n)" before it if the current expression is an
ObjectIntersectionOf.

This allows correctly translating an expression of the form "X and Y
that P some Z".