Allow using a custom connecting word with some object properties.
Previously all existential restriction expressions (P some X) were
rendered with the connecting word "some" between the property and the
filler expression, *except* when the filler expression was a FlyBase
gene (typically used when the property is RO:0002292, "expresses").

We introduced a generalized version that allows overriding the
connecting word for any object property, at the same time we override
the property's label.

We also add a few more properties with an overriden label.